+++ recipes/ruby.rb
@@ -13,7 +13,11 @@ cacert_file = ENV[OpenSSL::X509::DEFAULT_CERT_FILE_ENV] || OpenSSL::X509::DEFAUL
 
 node['doi_ssl_filtering']['cert_locations'].each do |loc|
   filename = get_cert_filemame(loc)
-  original_file = File.join(Chef::Config[:file_cache_path], filename)
+  original_file = if platform_family?('windows')
+                    File.join(Chef::Config[:file_cache_path], filename).gsub(File::ALT_SEPARATOR,File::SEPARATOR)
+                  else
+                    File.join(Chef::Config[:file_cache_path], filename)
+                  end
   local_ssl_file = File.join(ca_path, filename)
 
   remote_file local_ssl_file do