+++ test/integration/windefault/serverspec/shared_tests_spec.rb
@@ -1,4 +1,0 @@
-require 'spec_helper.rb'
-
-require_relative '../shared_tests/certificate_spec.rb'
-require_relative '../shared_tests/kitchen_spec.rb'