+++ metadata.rb
@@ -6,9 +6,10 @@ description 'Installs/Configures openssl and ruby applications for DOI SSL filte
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 issues_url 'https://github.com/USGS-CIDA/chef-cookbook-doi-ssl-filtering/issues'
 source_url 'https://github.com/USGS-CIDA/chef-cookbook-doi-ssl-filtering'
-version '1.0.6'
+version '1.0.7'
 
 supports 'centos', '>= 6.8'
 supports 'centos', '>= 7.0'
 chef_version '>= 12.0.0'
 depends 'java'
+depends 'windows'