+++ test/integration/java-oracle/serverspec/java_spec.rb
@@ -1,3 +1,0 @@
-require 'spec_helper.rb'
-
-require_relative '../../suites/shared_tests/java_spec.rb'