+++ test/integration/java10-oracle/serverspec/kitchen_spec.rb
@@ -1,3 +1,0 @@
-require 'spec_helper.rb'
-
-require_relative '../../suites/shared_tests/kitchen_spec.rb'